+++ superset/dashboards/commands/export.py
@@ -33,7 +33,8 @@ from superset.datasets.commands.export import ExportDatasetsCommand
 from superset.datasets.dao import DatasetDAO
 from superset.models.dashboard import Dashboard
 from superset.models.slice import Slice
-from superset.utils.dict_import_export import EXPORT_VERSION, get_filename
+from superset.utils.dict_import_export import EXPORT_VERSION
+from superset.utils.file import get_filename
 
 logger = logging.getLogger(__name__)
