+++ tests/unit_tests/utils/test_file.py
@@ -17,7 +17,7 @@
 # under the License.
 import pytest
 
-from superset.utils.import_export import get_filename
+from superset.utils.file import get_filename
 
 
 @pytest.mark.parametrize(
@@ -29,8 +29,12 @@ from superset.utils.import_export import get_filename
         ("D:\\Charts\\Energy Sankey", 132, True, "DChartsEnergy_Sankey"),
         ("🥴🥴🥴", 4751, False, "4751"),
         ("🥴🥴🥴", 4751, True, "4751"),
+        ("Energy Sankey 🥴🥴🥴", 4751, False, "Energy_Sankey_4751"),
+        ("Energy Sankey 🥴🥴🥴", 4751, True, "Energy_Sankey"),
         ("你好", 475, False, "475"),
         ("你好", 475, True, "475"),
+        ("Energy Sankey 你好", 475, False, "Energy_Sankey_475"),
+        ("Energy Sankey 你好", 475, True, "Energy_Sankey"),
     ],
 )
 def test_get_filename(
